Update all page.tsx, rendering directly from server insteand of "use client", also update dashboard to allow user back to home page
