+++ src/app/(auth)/login/components/DecorativeElements.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Image from 'next/image';
 import React from 'react';
